+++ dotcom-rendering/src/components/RichLinkComponent.importable.tsx
@@ -1,3 +1,4 @@
+import type { FEFormat } from '../frontend/feArticle';
 import { type ArticleFormat, decideFormat } from '../lib/articleFormat';
 import { useApi } from '../lib/useApi';
 import type {
@@ -5,7 +6,6 @@ import type {
 	RichLinkBlockElement,
 	StarRating,
 } from '../types/content';
-import type { FEFormat } from '../types/frontend';
 import type { RichLinkCardType } from '../types/layout';
 import type { TagType } from '../types/tag';
 import { RichLink } from './RichLink';
