+++ dotcom-rendering/src/server/handler.article.amp.tsx
@@ -11,7 +11,7 @@ import { generatePermutivePayload } from '../lib/permutive.amp';
 import { extractScripts } from '../lib/scripts.amp';
 import { findBySubsection } from '../model/article-sections';
 import { extractNAV } from '../model/extract-nav';
-import { validateAsArticleType } from '../model/validate';
+import { validateAsFEArticle } from '../model/validate';
 import { getAmpExperimentCache } from './AMPExperimentCache.amp';
 import { recordTypeAndPlatform } from './lib/logging-store';
 import { renderArticle } from './render.article.amp';
@@ -19,7 +19,7 @@ import { renderArticle } from './render.article.amp';
 export const handleAMPArticle: RequestHandler = ({ body }, res, next) => {
 	(async () => {
 		recordTypeAndPlatform('article', 'amp');
-		const article = validateAsArticleType(body);
+		const article = validateAsFEArticle(body);
 		const format = decideFormat(article.format);
 
 		if (
