+++ dotcom-rendering/src/server/handler.article.apps.ts
@@ -1,7 +1,7 @@
 import type { RequestHandler } from 'express';
 import { decideFormat } from '../lib/articleFormat';
 import { enhanceBlocks } from '../model/enhanceBlocks';
-import { validateAsArticleType, validateAsBlock } from '../model/validate';
+import { validateAsBlock, validateAsFEArticle } from '../model/validate';
 import { enhanceArticleType } from '../types/article';
 import type { FEBlocksRequest } from '../types/frontend';
 import { makePrefetchHeader } from './lib/header';
@@ -11,7 +11,7 @@ import { renderAppsBlocks, renderArticle } from './render.article.apps';
 export const handleAppsArticle: RequestHandler = ({ body }, res) => {
 	recordTypeAndPlatform('article', 'apps');
 
-	const frontendData = validateAsArticleType(body);
+	const frontendData = validateAsFEArticle(body);
 	const article = enhanceArticleType(frontendData, 'Apps');
 	const { html, prefetchScripts } = renderArticle(article);
 
@@ -22,7 +22,7 @@ export const handleAppsArticle: RequestHandler = ({ body }, res) => {
 export const handleAppsInteractive: RequestHandler = ({ body }, res) => {
 	recordTypeAndPlatform('interactive', 'app');
 
-	const frontendData = validateAsArticleType(body);
+	const frontendData = validateAsFEArticle(body);
 	const article = enhanceArticleType(frontendData, 'Apps');
 	const { html, prefetchScripts } = renderArticle(article);
