+++ dotcom-rendering/src/types/article.amp.tsx
@@ -1,8 +1,9 @@
+import type { FEFormat } from '../frontend/feArticle';
 import type { EditionId } from '../lib/edition';
 import type { Block } from './blocks';
 import type { CommercialProperties } from './commercial';
 import type { FEElement } from './content';
-import type { FEFormat, FELinkType, LegacyPillar } from './frontend';
+import type { FELinkType, LegacyPillar } from './frontend';
 import type { PaginationType } from './liveBlog';
 import type { TagType } from './tag';
